requirements.txt
six
Faker
pillow
markdown
python-dateutil

django
django-environ
django-filter
django-allauth
django-resized
python-decouple
django-ckeditor
django-browser-reload
django-crispy-forms
crispy-bootstrap5

# PAYMENTS
django-payments
django-payments[stripe]